Added sorting for STH ymls
All lines will now be sorted alphabetically before they are put in the STH yml files. This is to compensate the localization sheets revamp changes, which enables the users to sort and filter columns. If the order on the sheets changes it is difficult to keep track of what has been actually changed (the diff gets very big as lines change positions)